+++ defender/downloaded/Contract_cToken.js
@@ -21,7 +21,7 @@ const saiTokenAddress = '0x89d24A6b4CcB1B6fAA2625fE562bDD9a23260359'.toLowerCase
 const oddTokens = [makerTokenAddress, saiTokenAddress];
 const cEtherAddress = '0x4Ddc2D193948926D02f9B1fE9e1daa0718270ED5'.toLowerCase();
 
-// Temporary to account for Proposal 117 
+// Temporary to account for Proposal 117
 const oldOracleAddress = '0x65c816077C29b557BEE980ae3cC2dCE80204A0C5';
 
 const eventMapping = {
@@ -369,7 +369,7 @@ exports.handler = async function (autotaskEvent) {
 
   // create an ethers.js Contract for the Compound Oracle contract
   const oracleContract = await getOracleContract(provider);
-  
+
   // Temporary due to Proposal 117
   const oldOracleContract = new ethers.Contract(
     oldOracleAddress,
@@ -385,7 +385,7 @@ exports.handler = async function (autotaskEvent) {
     // to make that determination
     const cTokenAddress = getAddressForMatchReason(reason, logs, abi);
     if (cTokenAddress === undefined) {
-        throw new Error('unable to get address for match reason');
+      throw new Error('unable to get address for match reason');
     }
 
     // determine the type of event it was
@@ -405,7 +405,7 @@ exports.handler = async function (autotaskEvent) {
         usdPerTokenDecimals,
       } = await getTokenPrice(oracleContract, cTokenAddress, decimals));
     } catch (error) {
-      console.debug(`Error using oracle, falling back to old oracle`);
+      console.debug('Error using oracle, falling back to old oracle');
       if (cTokenAddress.toLowerCase() !== cEtherAddress) {
         console.error(`cToken address is not cEther: ${cTokenAddress}`);
         throw error;
@@ -435,8 +435,8 @@ exports.handler = async function (autotaskEvent) {
   const etherscanLink = `[TX](<https://etherscan.io/tx/${transactionHash}>)`;
 
   const discordPromises = messages.map((message) => {
-      console.log(`${etherscanLink} ${message}`);
-      return postToDiscord(discordUrl, `${etherscanLink} ${message}`);
+    console.log(`${etherscanLink} ${message}`);
+    return postToDiscord(discordUrl, `${etherscanLink} ${message}`);
   });
 
   // wait for the promises to settle
