+++ defender/tests/Forta_Large_Delegations.spec.js
@@ -1,161 +1,149 @@
-jest.mock('axios', () => jest.fn());
+// Set the name of the Secret set in Autotask
+const discordSecretName = 'SecurityAlertsDiscordUrl';
+// Name of the Secret in the .env file
+const discordEnvSecretName = 'discordUrl';
+
+// Mock the data from the Bot finding
+const mockTxHash = '0x2c9931793876db33b1a9aad123ad4921dfb9cd5e59dbb78ce78f277759587115';
+const mockBlockHash = '0xfc492bb1149eaad9dbffb1990003ce203a9c8102fb57cd767b45c6c4749c86c5';
+const mockMetadata = {
+  delegateAddress: '0x1212121212121212121212121212121212121212',
+  levelName: 'proposal',
+  minAmountCOMP: '11',
+  delegateCOMPBalance: '12',
+};
+
+// mock the axios package
+const acceptedPost = {
+  status: 204,
+  statusText: 'No Content',
+};
+jest.mock('axios', () => jest.fn().mockResolvedValue(acceptedPost));
+// eslint-disable-next-line import/no-extraneous-dependencies
 const axios = require('axios');
 
-const { handler } = require('../downloaded/Forta_Large_Delegations');
+const {
+  Finding, FindingType, FindingSeverity,
+} = require('forta-agent');
+
+const mockFinding = Finding.fromObject({
+  name: 'Placeholder Alert',
+  description: 'Placeholder description',
+  alertId: 'AE-ALERT-ID',
+  type: FindingType.Info,
+  severity: FindingSeverity.Info,
+  protocol: 'Protocol',
+  metadata: mockMetadata,
+});
 
 // grab the existing keys before loading new content from the .env file
 const existingKeys = Object.keys(process.env);
+// eslint-disable-next-line import/no-unresolved
 require('dotenv').config();
 
 // now filter out all of the existing keys from what is currently in the process.env Object
-const newKeys = Object.keys(process.env).filter((key) => !existingKeys.includes(key));
+const newKeys = Object.keys(process.env).filter((key) => existingKeys.indexOf(key) === -1);
 const secrets = {};
 newKeys.forEach((key) => {
   secrets[key] = process.env[key];
 });
 
-const transactionHash = '0x9876543210FEDCBA';
-const delegateAddress = '0x1212121212121212121212121212121212121212';
-
-const mockFortaAlert = {
-  data: {
-    data: {
-      alerts: {
-        pageInfo: {
-          hasNextPage: false,
-          endCursor: {
-            alertId: 'AE-COMP-GOVERNANCE-DELEGATE-THRESHOLD',
-            blockNumber: 0,
-          },
-        },
-        alerts: [
-          {
-            createdAt: '2022-03-31T22:02:20.812799122Z',
-            name: 'Compound Governance Delegate Threshold Alert',
-            protocol: 'Compound',
-            findingType: 'Suspicious',
-            hash: '0xcee8d4bd1c065260acdcfa51c955fc29c984145de2769b685f29701b6edf318f',
-            alertId: 'AE-COMP-GOVERNANCE-DELEGATE-THRESHOLD',
-            source: {
-              transactionHash,
-              block: {
-                number: 14496506,
-                chainId: 1,
-              },
-              bot: {
-                id: '0x3f02bee8b17edc945c5c1438015aede79225ac69c46e9cd6cff679bb71f35576',
-              },
-            },
-            severity: 'Medium',
-            metadata: {
-              delegateAddress,
-              levelName: 'proposal',
-              minAmountCOMP: '11',
-              delegateCOMPBalance: '45',
-            },
-            description: `The address ${delegateAddress} has been delegated enough COMP token to pass the minimum threshold for the governance event: proposal`,
-          },
-        ],
-      },
-    },
-  },
-};
-
-// create a provider that will be injected as the Defender Relayer provider
-jest.mock('defender-relay-client/lib/ethers', () => ({
-  DefenderRelayProvider: jest.fn(),
-}));
-
-function createFortaSentinelEvents(botId) {
-  const { alerts } = mockFortaAlert.data.data.alerts;
-  const autotaskEvents = alerts.map((alert) => {
-    // augment the alert Object with additional fields
-    // admittedly, there is some hand-waving here because we have to mock some of the Sentinel
-    // fields that don't originate from the Forta Public API
-    // e.g. We have to specify the alertId in the Sentinel to perform filtering on what we get from
-    // the Forta Bot in the first place.
-    /* eslint-disable no-param-reassign */
-    alert.source.bot.name = 'N/A';
-    alert.source.block.chain_id = alert.source.block.chainId;
-    alert.source.tx_hash = alert.source.transactionHash;
-    alert.alertType = 'TX';
-    alert.alert_id = 'ALERT_ID_PLACEHOLDER';
-    alert.type = 'INFORMATION';
-    alert.scanner_count = 1;
-    /* eslint-enable no-param-reassign */
-
-    // populate the matchReasons Array with placeholders
-    const matchReasons = [
-      {
-        type: 'alert-id',
-        value: 'ALERT_ID_PLACEHOLDER',
-      },
-      {
-        type: 'severity',
-        value: 'INFO',
-      },
-    ];
-
-    // populate the sentinel Object with placeholders
-    // none of these are currently checked by any Autotasks in use
-    const sentinel = {
-      id: '8fe3d50b-9b52-44ff-b3fd-a304c66e1e56',
-      name: 'Sentinel Name Placeholder',
-      addresses: [],
-      bots: [botId],
-      network: 'mainnet',
-      chainId: 1,
-    };
-
-    const autotaskEvent = {
-      relayerARN: undefined,
-      kvstoreARN: undefined,
-      credentials: undefined,
-      tenantId: undefined,
-      secrets,
-      request: {
-        body: {
-          hash: alert.hash, // forta Bot hash
-          alert,
-          matchReasons,
-          sentinel,
-          type: 'FORTA',
-        },
-      },
-    };
-    return autotaskEvent;
-  });
-
-  return autotaskEvents;
+// Map the Env name to the Secret variable name
+if (discordSecretName !== discordEnvSecretName) {
+  secrets[discordSecretName] = secrets[discordEnvSecretName];
+  delete secrets[discordEnvSecretName];
 }
 
-it('Runs autotask against blocks in configuration file', async () => {
-  const results = [];
-  const mockBotId = '0x12345';
-  const autotaskEvents = createFortaSentinelEvents(mockBotId);
+// eslint-disable-next-line import/no-useless-path-segments
+const { handler } = require('../downloaded/Forta_Large_Delegations');
 
-  // pass the mocked Forta Bot alert into the function that will emulate a Forta Sentinel alert
-  // update the axios mock in preparation for capturing the Discord POST request
-  axios.post = jest.fn().mockResolvedValue(mockFortaAlert);
-  axios.mockImplementation((arg0) => { results.push(arg0); });
+function createFortaSentinelEvent(finding, blockHash, txHash) {
+  // Generally findings go from the Bot, to Scan Node, to Sentinel, to Autotasks
+  //  with some metadata being added and removed along the way. This function will mimic
+  // the Sentinel output with only Finding, block and transaction data.
 
-  // run the autotask on the events
-  const promises = autotaskEvents.map((autotaskEvent) => handler(autotaskEvent));
+  // Note: Much of the extra data here is superfluous but is left here just in case future bots
+  // want to reference any of the Sentinel data in the Discord output. It also mimics sentinel
+  // output more accurately.
 
-  await Promise.all(promises);
-
-  const expectedResults = [
+  // populate the matchReasons Array with placeholders
+  const matchReasons = [
     {
-      url: '',
-      method: 'post',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      data: {
-        content: `[TX](<https://etherscan.io/tx/${transactionHash}>) 💸 **0x1212** has had enough **COMP** tokens delegated to pass min threshold for the governance event: **proposal**`,
-      },
+      type: 'alert-id',
+      value: 'ALERT_ID_PLACEHOLDER',
+    },
+    {
+      type: 'severity',
+      value: 'INFO',
     },
   ];
 
-  expect(results).toStrictEqual(expectedResults);
+  // populate the sentinel Object with placeholders
+  const sentinel = {
+    id: '8fe3d50b-9b52-44ff-b3fd-a304c66e1e56',
+    name: 'Sentinel Name Placeholder',
+    addresses: [],
+    agents: [],
+    network: 'mainnet',
+    chainId: 1,
+  };
+
+  const autotaskEvent = {
+    relayerARN: undefined,
+    kvstoreARN: undefined,
+    credentials: undefined,
+    tenantId: undefined,
+    secrets,
+    request: {
+      body: {
+        hash: '0xAGENT-HASH', // forta Agent hash
+        alert: {
+          metadata: finding.metadata,
+        },
+        source: {
+          transactionHash: txHash,
+          block: {
+            hash: blockHash,
+          },
+        },
+        matchReasons,
+        sentinel,
+        type: 'FORTA',
+      },
+    },
+  };
+  return autotaskEvent;
+}
+
+describe('check autotask', () => {
+  const url = secrets[discordSecretName];
+  const headers = { 'Content-Type': 'application/json' };
+  const method = 'post';
+
+  beforeEach(async () => {
+    axios.mockClear();
+  });
+
+  it('Runs autotask against mock data and posts in Discord', async () => {
+    const autotaskEvent = createFortaSentinelEvent(mockFinding, mockBlockHash, mockTxHash);
+    // run the autotask on the events
+    await handler(autotaskEvent);
+
+    const data = '{"content":"[TX](<https://etherscan.io/tx/0x2c9931793876db33b1a9aad123ad4921dfb9cd5e59dbb78ce78f277759587115>) 💸 **0x1212** has been delegated  enough **COMP** tokens to pass min threshold for the governance event: **proposal**"}';
+    const expectedLastCall = {
+      url, headers, method, data,
+    };
+    expect(axios).toBeCalledTimes(1);
+    expect(axios.mock.lastCall[0]).toStrictEqual(expectedLastCall);
+  });
+
+  it('throws error if discordUrl is not valid', async () => {
+    // Use an invalid discord URL
+    secrets[discordSecretName] = 'http//zzzz';
+    const autotaskEvent = createFortaSentinelEvent(mockFinding, mockBlockHash, mockTxHash);
+    // run the autotask on the events
+    await expect(handler(autotaskEvent)).rejects.toThrow('discordUrl is not a valid URL');
+
+    expect(axios).toBeCalledTimes(0);
+  });
 });
