+++ defender/tests/Forta_Liquidation_Monitor.spec.js
@@ -1,5 +1,39 @@
-jest.mock('axios', () => jest.fn());
+// Set the name of the Secret set in Autotask
+const discordSecretName = 'LiquidationDiscordUrl';
+// Name of the Secret in the .env file
+const discordEnvSecretName = 'discordUrl';
+
+// Mock the data from the Bot finding
+const mockBlockHash = '0x1110890564dbd87ca848b7107487ae5a7d28da1b16707bccd3ba37381ae33419';
+const mockMetadata = {
+  borrowerAddress: '0x0000000000000000000000000000000000000000',
+  liquidationAmount: '1000.00',
+  shortfallAmount: '1000.00',
+  healthFactor: '0.80',
+};
+
+// mock the axios package
+const acceptedPost = {
+  status: 204,
+  statusText: 'No Content',
+};
+jest.mock('axios', () => jest.fn().mockResolvedValue(acceptedPost));
+// eslint-disable-next-line import/no-extraneous-dependencies
 const axios = require('axios');
+
+const {
+  Finding, FindingType, FindingSeverity,
+} = require('forta-agent');
+
+const mockFinding = Finding.fromObject({
+  name: 'Placeholder Alert',
+  description: 'Placeholder description',
+  alertId: 'AE-ALERT-ID',
+  type: FindingType.Info,
+  severity: FindingSeverity.Info,
+  protocol: 'Protocol',
+  metadata: mockMetadata,
+});
 
 // grab the existing keys before loading new content from the .env file
 const existingKeys = Object.keys(process.env);
@@ -13,58 +47,16 @@ newKeys.forEach((key) => {
   secrets[key] = process.env[key];
 });
 
-const { handler } = require('../downloaded/Forta_Liquidation_Monitor');
+// Map the Env name to the Secret variable name
+if (discordSecretName !== discordEnvSecretName) {
+  secrets[discordSecretName] = secrets[discordEnvSecretName];
+  delete secrets[discordEnvSecretName];
+}
 
-const blockHash = '0xFAKEBLOCKHASH';
-const transactionHash = '0xb85bbcdfd06edf6bcaa3271e49a339cc878daa30ec5f987a43a4d11d925ba751';
-const borrowerAddress = '0x0000000000000000000000000000000000000000';
+// eslint-disable-next-line import/no-useless-path-segments
+const { handler } = require('../staged/downloaded/Forta_Liquidation_Monitor');
 
-const mockFortaAlert = {
-  data: {
-    data: {
-      alerts: {
-        pageInfo: {
-          hasNextPage: false,
-          endCursor: {
-            alertId: 'AE-COMP-LIQUIDATION-THRESHOLD',
-            blockNumber: 0,
-          },
-        },
-        alerts: [
-          {
-            createdAt: '2022-03-31T22:02:20.812799122Z',
-            name: 'Compound Liquidation Threshold Alert',
-            protocol: 'Compound',
-            findingType: 'Info',
-            hash: '0xcee8d4bd1c065260acdcfa51c955fc29c984145de2769b685f29701b6edf318f',
-            alertId: 'AE-COMP-LIQUIDATION-THRESHOLD',
-            source: {
-              transactionHash,
-              block: {
-                number: 14496506,
-                chainId: 1,
-                hash: blockHash,
-              },
-              bot: {
-                id: '0x3f02bee8b17edc945c5c1438015aede79225ac69c46e9cd6cff679bb71f35576',
-              },
-            },
-            severity: 'Info',
-            metadata: {
-              borrowerAddress,
-              liquidationAmount: '1000.00',
-              shortfallAmount: '1000.00',
-              healthFactor: '0.80',
-            },
-            description: 'The address 0x1111 has dropped below the liquidation threshold. The account may be liquidated for: $515.15 USD',
-          },
-        ],
-      },
-    },
-  },
-};
-
-function createFortaSentinelEvents(botId) {
+function createFortaSentinelEvent(finding, blockHash) {
   // Generally findings go from the Bot, to Scan Node, to Sentinel, to Autotasks
   //  with some metadata being added and removed along the way. This function will mimic
   // the Sentinel output with only Finding, block and transaction data.
@@ -72,104 +64,90 @@ function createFortaSentinelEvents(botId) {
   // Note: Much of the extra data here is superfluous but is left here just in case future bots
   // want to reference any of the Sentinel data in the Discord output. It also mimics sentinel
   // output more accurately.
-  const { alerts } = mockFortaAlert.data.data.alerts;
-  const autotaskEvents = alerts.map((alert) => {
-    // augment the alert Object with additional fields
-    // admittedly, there is some hand-waving here because we have to mock some of the Sentinel
-    // fields that don't originate from the Forta Public API
-    // e.g. We have to specify the alertId in the Sentinel to perform filtering on what we get from
-    // the Forta Bot in the first place.
-    /* eslint-disable no-param-reassign */
-    alert.source.bot.name = 'N/A';
-    alert.alert_id = alert.alertId;
-    alert.alertType = 'TX';
-    alert.type = 'INFORMATION';
-    alert.scanNodeCount = 1;
-    /* eslint-enable no-param-reassign */
 
-    // populate the matchReasons Array with placeholders
-    const matchReasons = [
-      {
-        type: 'alert-id',
-        value: 'ALERT_ID_PLACEHOLDER',
-      },
-      {
-        type: 'severity',
-        value: 'INFO',
-      },
-    ];
+  // On block events, the txHash does not exist
+  const txHash = '';
 
-    // populate the sentinel Object with placeholders
-    // none of these are currently checked by any Autotasks in use
-    const sentinel = {
-      id: '8fe3d50b-9b52-44ff-b3fd-a304c66e1e56',
-      name: 'Sentinel Name Placeholder',
-      addresses: [],
-      bots: [botId],
-      network: 'mainnet',
-      chainId: 1,
-    };
+  // populate the matchReasons Array with placeholders
+  const matchReasons = [
+    {
+      type: 'alert-id',
+      value: 'ALERT_ID_PLACEHOLDER',
+    },
+    {
+      type: 'severity',
+      value: 'INFO',
+    },
+  ];
 
-    const autotaskEvent = {
-      relayerARN: undefined,
-      kvstoreARN: undefined,
-      credentials: undefined,
-      tenantId: undefined,
-      secrets,
-      request: {
-        body: {
-          hash: alert.hash, // forta Bot hash
-          alert,
-          matchReasons,
-          sentinel,
-          type: 'FORTA',
+  // populate the sentinel Object with placeholders
+  const sentinel = {
+    id: '8fe3d50b-9b52-44ff-b3fd-a304c66e1e56',
+    name: 'Sentinel Name Placeholder',
+    addresses: [],
+    agents: [],
+    network: 'mainnet',
+    chainId: 1,
+  };
+
+  const autotaskEvent = {
+    relayerARN: undefined,
+    kvstoreARN: undefined,
+    credentials: undefined,
+    tenantId: undefined,
+    secrets,
+    request: {
+      body: {
+        hash: '0xAGENT-HASH', // forta Agent hash
+        alert: {
+          metadata: finding.metadata,
         },
+        source: {
+          transactionHash: txHash,
+          block: {
+            hash: blockHash,
+          },
+        },
+        matchReasons,
+        sentinel,
+        type: 'FORTA',
       },
-    };
-    return autotaskEvent;
-  });
-
-  return autotaskEvents;
+    },
+  };
+  return autotaskEvent;
 }
 
 describe('check autotask', () => {
-  it('Runs autotask against mock data and posts in Discord (manual-check)', async () => {
-    const results = [];
-    const mockBotId = '0x12345';
-    const autotaskEvents = createFortaSentinelEvents(mockBotId);
+  const url = secrets[discordSecretName];
+  const headers = { 'Content-Type': 'application/json' };
+  const method = 'post';
 
-    // pass the mocked Forta Bot alert into the function that will emulate a Forta Sentinel alert
-    // update the axios mock in preparation for capturing the Discord POST request
-    axios.mockImplementationOnce((arg0) => results.push(arg0));
+  beforeEach(async () => {
+    axios.mockClear();
+  });
+
+  it('Runs autotask against mock data and posts in Discord', async () => {
+    const autotaskEvent = createFortaSentinelEvent(mockFinding, mockBlockHash);
 
     // run the autotask on the events
-    const promises = autotaskEvents.map((autotaskEvent) => handler(autotaskEvent));
+    await handler(autotaskEvent);
 
-    await Promise.all(promises);
-
-    const expectedResults = [
-      {
-        url: 'https://discord.com/api/webhooks/963891500897423360/Zx8MzPcEfFPDoqpOjGXoBu3303FPvj0NAX4pHUsOll3G5N2TlaThiQUOUDfyQm0tWhiP',
-        method: 'post',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        data: {
-          content: `[BLOCK](<https://etherscan.io/block/${blockHash}>) - [ACCT](<https://etherscan.io/address/${borrowerAddress}>) 📉💵🔥 **Liquidatable account detected** account 0x0000 is liquidatable for $1000.00`,
-        },
-      },
-    ];
-
-    expect(results).toStrictEqual(expectedResults);
+    const data = '{"content":"[BLOCK](<https://etherscan.io/block/0x1110890564dbd87ca848b7107487ae5a7d28da1b16707bccd3ba37381ae33419>) - [ACCT](<https://etherscan.io/address/0x0000000000000000000000000000000000000000>) 📉💵🔥 **Liquidatable account detected** account 0x0000 is liquidatable for $1000.00"}';
+    const expectedLastCall = {
+      url, headers, method, data,
+    };
+    expect(axios).toBeCalledTimes(1);
+    expect(axios.mock.lastCall[0]).toStrictEqual(expectedLastCall);
   });
 
   it('throws error if discordUrl is not valid', async () => {
     // Use an invalid discord URL
-    secrets.discordUrl = 'http//zzzz';
-    const mockBotId = '0x12345';
-    const autotaskEvents = createFortaSentinelEvents(mockBotId);
+    secrets[discordSecretName] = 'http//zzzz';
+    const autotaskEvent = createFortaSentinelEvent(mockFinding, mockBlockHash);
 
     // run the autotask on the events
-    await expect(handler(autotaskEvents[0])).rejects.toEqual(new Error('discordUrl is not a valid URL'));
+    await expect(handler(autotaskEvent)).rejects.toThrow('discordUrl is not a valid URL');
+
+    expect(axios).toBeCalledTimes(0);
   });
 });
